# Improve HackAndSlash, RankingBattle execution time