`Lib9c.Benchmarks` result from latest 20 blocks.

before

```
Loading blocks  00:00:03.4116559
Executing actions       6438ms
Average per block       306ms
Average per tx  90ms
Average per action      76ms
Total elapsed   00:00:09.8500511
```

after

```
Loading blocks  00:00:02.6202123
Executing actions       5321ms
Average per block       253ms
Average per tx  74ms
Average per action      63ms
Total elapsed   00:00:07.9416374
```